@@ -1,41 +1,36 @@
+
 {
-  "name": "au-pair-frontend",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@radix-ui/react-label": "^2.1.7",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@supabase/supabase-js": "^2.50.3",
-    "@types/node": "^20.0.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "autoprefixer": "^10.0.0",
-    "axios": "^1.6.0",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.0",
-    "framer-motion": "^12.23.0",
-    "lucide-react": "^0.344.0",
-    "next": "^15.3.5",
-    "next-themes": "^0.4.6",
-    "postcss": "^8.0.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-select": "^5.10.1",
-    "socket.io-client": "^4.7.0",
-    "tailwind-merge": "^2.2.0",
-    "tailwindcss": "^3.4.0",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.0.0",
-    "zustand": "^4.4.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "next": "15.3.4",
+    "axios": "^1.6.2",
+    "socket.io-client": "^4.7.4",
+    "react-hook-form": "^7.48.2",
+    "date-fns": "^3.0.0",
+    "js-cookie": "^3.0.5",
+    "clsx": "^2.0.0",
+    "tailwind-merge": "^2.2.0"
   },
   "devDependencies": {
-    "eslint": "^8.0.0",
-    "eslint-config-next": "15.1.5"
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "@types/js-cookie": "^3.0.6",
+    "@tailwindcss/postcss": "^4",
+    "tailwindcss": "^4",
+    "eslint": "^9",
+    "eslint-config-next": "15.3.4",
+    "@eslint/eslintrc": "^3"
   }
 }
